feat(task-6): add event emitter to the course form - fix delete course handler - add course form event listeners

+++ src/app/modules/course/components/courses-list/courses-list.component.ts
@@ -68,8 +68,10 @@ export class CoursesListComponent implements OnInit {
     });
 
     dialogRef.afterClosed().subscribe((result: boolean) => {
-      this.courses$ = this.courseService.removeCourse(id);
-      this.ref.markForCheck();
+      if (result) {
+        this.courses$ = this.courseService.removeCourse(id);
+        this.ref.markForCheck();
+      }
     });
   }
 }
